backend/requirements.txt
fastapi
uvicorn
python-multipart
paddleocr
paddlepaddle
pillow
pydantic
sqlalchemy
python-dotenv
ollama
loguru